\name{geecure-package}
\alias{geecure-package}
\docType{package}
\title{Marginal proportional hazards mixture cure models with generalzied estimating equations}
\description{A package that uses generalized estimating equations (GEE) approach to estimate marginal proportional hazards mixture cure (PHMC) models. This package implements recently developed inference procedures for the marginal PHMC models with the expectation-solution (ES) algorithm. The package includes the parametric PHMC model with Weibull baseline distribution in the latency part and the semiparametric PHMC model for fitting the multivariate survival data with a cure fraction.}
\details{
{Package: geecure \cr
 Type: Package \cr
 Version: 1.0-2  \cr
 Date: 2017-11-26 \cr
 License: GPL(>=2) \cr
 LazyData: TRUE 
}}
\author{
Yi Niu <niuyi@dlut.edu.cn>, Hui Song, Xiaoguang Wang, Yingwei Peng
%Maintainer: Yi Niu \email{niuyi@dlut.edu.cn}
}
\references{
Liang, K.-Y. and Zeger, S. L. (1986) Longitudinal data analysis using generalized linear models. \emph{Biometrika}, \bold{73}: 13-22.

Niu, Y. and Peng, Y. (2013) A semiparametric marginal mixture cure model for clustered survival data. \emph{Statistics in Medicine}, \bold{32}: 2364-2373.

Niu, Y. and Peng, Y. (2014) Marginal regression analysis of clustered failure time data with a cure fraction. \emph{Journal of Multivariate Analysis}, \bold{123}: 129-142.

Niu, Y., Song, L., Liu, Y, and Peng, Y. (2017) Modeling clustered long-term survivors using marginal mixture cure model. under revision.

Peng, Y., Taylor, J. M. G, and Yu, B. (2007) A marginal regression model for multivariate failure time data with a surviving fraction. \emph{Lifetime Data Analysis}, \bold{13}: 351-369

Rosen, O., Jiang, W., and Tanner, M. A. (2000) Mixtures of marginal models. \emph{Biometrika}, \bold{87}: 391-404.

Yu, B. and Peng, Y. (2008) Mixture cure models for multivariate survival data. \emph{Computational Statistics & Data Analysis}, \bold{52}: 1524-1532.
}