\name{print.geecure}
\alias{print.geecure}
\title{Print geecure object}
\description{Output of \code{geecure} object.}
\usage{
%print.geecure(x, ...)
\method{print}{geecure}(x, ...)
}
\arguments{
  \item{x}{an object of \code{geecure}.}
  \item{\dots}{further arguments to be added in the \code{print.geecure} function.}
}