\name{print.geecure2}
\alias{print.geecure2}
\title{Print geecure2 object}
\description{Output of \code{geecure2} object.}
\usage{
%print.geecure2(x, ...)
\method{print}{geecure2}(x, ...)
}
\arguments{
  \item{x}{an object of \code{geecure2}.}
  \item{\dots}{further arguments to be added in the \code{print.geecure2} function.}
}